Fix issue #59

+++ app/src/main/java/solutions/s4y/itag/MainActivity.java
@@ -228,7 +228,7 @@ public class MainActivity extends FragmentActivity implements LeScanner.LeScanne
         if (hasFocus!=mHasFocus) {
             mHasFocus = hasFocus;
             if (mIsServiceStartedUnbind) {
-                if (mHasFocus) {
+                if (mHasFocus && mITagsService!=null) {
                     mITagsService.removeFromForeground();
                 }
             }
